Backend/PND.AI/requirements.txt
fastapi
uvicorn
tensorflow-cpu
Pillow
numpy
python-multipart